Change AddStatement class to local class
AddStatement 클래스를 add 메소드의 로컬 클래스로 변경
내부 클래스 이므로 로컬 클래스에서 add 메소드의 로컬 변수를 직접 참조할 수 있음(단, final 혹은 effectively final 이어야 함)
로컬 클래스를 사용하므로 클래스 파일 개수를 줄일 수 있음
