--- src/main/java/com/example/tobyspring/user/dao/AddStatement.java
@@ -1,26 +1,0 @@
-package com.example.tobyspring.user.dao;
-
-import com.example.tobyspring.user.domain.User;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-public class AddStatement implements StatementStrategy {
-    private User user;
-
-    public AddStatement(User user) {
-        this.user = user;
-    }
-
-    @Override
-    public PreparedStatement makePreparedStatement(Connection c) throws SQLException {
-        PreparedStatement ps = c.prepareStatement("insert into users(id, name, password) values(?,?,?)");
-
-        ps.setString(1, user.getId());
-        ps.setString(2, user.getName());
-        ps.setString(3, user.getPassword());
-
-        return ps;
-    }
-}
