+++ src/main/java/com/example/tobyspring/user/dao/UserDao.java
@@ -20,7 +20,20 @@ public class UserDao {
 */
 
     public void add(User user) throws SQLException {
-        StatementStrategy st = new AddStatement(user);
+        class AddStatement implements StatementStrategy {
+            @Override
+            public PreparedStatement makePreparedStatement(Connection c) throws SQLException {
+                PreparedStatement ps = c.prepareStatement("insert into users(id, name, password) values(?,?,?)");
+
+                ps.setString(1, user.getId());
+                ps.setString(2, user.getName());
+                ps.setString(3, user.getPassword());
+
+                return ps;
+            }
+        }
+
+        StatementStrategy st = new AddStatement();
         jdbcContextWithStatementStrategy(st);
     }
